# services/go/payment/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Payment Service API
  description: |
    NOWPayments integration API for crypto deposits and withdrawals on a gambling platform.
    
    ## Authentication
    All endpoints require Bearer token authentication via the `Authorization` header.
    
    ## Rate Limiting
    - 100 requests per minute per user
    - 1000 requests per minute per IP for webhook endpoints
    
    ## Idempotency
    All write operations support idempotency via the `idempotency_key` field.
    Replaying a request with the same key returns the original response.
    
    ## Error Codes
    Payment service errors use codes in the 5000-5999 range.
  version: 1.0.0
  contact:
    name: Platform Team
    email: platform@example.com
servers:
  - url: https://api.example.com
    description: Production
  - url: https://api-staging.example.com
    description: Staging
...
security:
  - BearerAuth: []
paths:
  /api/v1/payments/deposit:
    post:
      tags:
        - Deposits
      summary: Initiate a crypto deposit
      description: |
        Creates a new deposit payment request and returns a crypto address for the user to send funds to.
        
        The user must have a valid KYC level. Daily deposit limits are enforced based on KYC level:
        - Level 0: $500/day
        - Level 1: $2,000/day
        - Level 2: $10,000/day
        - Level 3: $50,000/day
      operationId: initiateDeposit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DepositRequest'
            examples:
              btc_deposit:
                summary: BTC Deposit
                value:
                  amount: "100.00"
                  currency: "BTC"
                  idempotency_key: "dep_abc123xyz789"
              usdt_deposit:
                summary: USDT TRC-20 Deposit
                value:
                  amount: "500.00"
                  currency: "USDTTRX"
                  idempotency_key: "dep_def456uvw012"
      responses:
        '201':
          description: Deposit created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepositResponse'
              example:
                success: true
                data:
                  payment_uuid: "550e8400-e29b-41d4-a716-446655440000"
                  payment_id: "123456789"
                  pay_address: "bc1qxy2kgdygjrsqtzq2n0yrf2493p83kkfjhx0wlh"
                  pay_amount: "0.0025"
                  pay_currency: "BTC"
                  fiat_amount: "100.00"
                  expires_at: "2024-01-15T12:00:00Z"
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalid_amount:
                  summary: Invalid amount format
                  value:
                    code: 4000
                    message: "invalid amount format"
                invalid_currency:
                  summary: Currency not supported
                  value:
                    code: 5009
                    message: "cryptocurrency not supported"
                    details:
                      currency: "DOGE"
        '403':
          description: KYC level insufficient
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 5004
                message: "KYC level 2 required for withdrawals"
                details:
                  current_level: 1
                  required_level: 2
        '422':
          description: Daily limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 5005
                message: "daily limit exceeded"
                details:
                  limit: 500.00
                  used: 450.00
                  requested: 100.00
                  available: 50.00
  /api/v1/payments/{id}:
    get:
      tags:
        - Deposits
      summary: Get payment by ID
      description: Retrieves a specific payment by its UUID
      operationId: getPayment
      parameters:
        - name: id
          in: path
          required: true
          description: Payment UUID
          schema:
            type: string
            format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '200':
          description: Payment found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
              example:
                success: true
                data:
                  uuid: "550e8400-e29b-41d4-a716-446655440000"
                  payment_id: "123456789"
                  user_id: 1001
                  requested_amount: "100.00"
                  actual_amount: "100.50"
                  fiat_amount: "100.00"
                  fiat_currency: "USD"
                  crypto_currency: "BTC"
                  pay_address: "bc1qxy2kgdygjrsqtzq2n0yrf2493p83kkfjhx0wlh"
                  pay_amount: "0.0025"
                  status: "finished"
                  created_at: "2024-01-15T10:00:00Z"
                  updated_at: "2024-01-15T10:30:00Z"
                  completed_at: "2024-01-15T10:30:00Z"
                  expires_at: "2024-01-15T12:00:00Z"
        '404':
          description: Payment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 5001
                message: "payment not found"
                details:
                  payment_id: "550e8400-e29b-41d4-a716-446655440000"
  /api/v1/payments/history:
    get:
      tags:
        - Deposits
      summary: List payment history
      description: Retrieves paginated payment history for the authenticated user
      operationId: getPaymentHistory
      parameters:
        - name: limit
          in: query
          description: Number of items per page (max 100)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          example: 20
        - name: cursor
          in: query
          description: Pagination cursor for next page
          schema:
            type: string
          example: "eyJpZCI6MTIzNDU2fQ=="
        - name: status
          in: query
          description: Filter by payment status
          schema:
            $ref: '#/components/schemas/PaymentStatus'
      responses:
        '200':
          description: Payment history retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentHistoryResponse'
              example:
                success: true
                data:
                  items:
                    - payment_uuid: "550e8400-e29b-41d4-a716-446655440000"
                      amount: "100.00"
                      currency: "BTC"
                      status: "finished"
                      created_at: "2024-01-15T10:00:00Z"
                    - payment_uuid: "660f9500-f30c-52e5-b827-557766551111"
                      amount: "50.00"
                      currency: "ETH"
                      status: "pending"
                      created_at: "2024-01-14T15:00:00Z"
                  next_cursor: "eyJpZCI6MTEwMDAwfQ=="
                  has_more: true
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from authentication service
  schemas:
...
    DepositRequest:
      type: object
      required:
        - amount
        - currency
        - idempotency_key
      properties:
        amount:
          type: string
          description: Fiat amount in USD (will be converted to crypto)
          pattern: '^\d+(\.\d{1,2})?$'
          example: "100.00"
        currency:
          type: string
          description: Cryptocurrency code
          enum:
            - BTC
            - ETH
            - USDTETH
            - USDTTRX
            - USDC
            - LTC
            - BCH
          example: "BTC"
        idempotency_key:
          type: string
          description: Unique key for idempotent requests
          minLength: 1
          maxLength: 255
          example: "dep_abc123xyz789"
...
    DepositResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            payment_uuid:
              type: string
              format: uuid
              description: Internal payment UUID
            payment_id:
              type: string
              description: NOWPayments payment ID
            pay_address:
              type: string
              description: Crypto address to send funds to
            pay_amount:
              type: string
              description: Expected crypto amount
            pay_currency:
              type: string
              description: Cryptocurrency code
            fiat_amount:
              type: string
              description: Fiat amount in USD
            expires_at:
              type: string
              format: date-time
              description: Payment expiration time
    PaymentResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/Payment'
    Payment:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        payment_id:
          type: string
        user_id:
          type: integer
          format: int64
        requested_amount:
          type: string
        actual_amount:
          type: string
          nullable: true
        fiat_amount:
          type: string
        fiat_currency:
          type: string
        crypto_currency:
          type: string
        pay_address:
          type: string
        pay_amount:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/PaymentStatus'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          nullable: true
        expires_at:
          type: string
          format: date-time
          nullable: true
    PaymentHistoryResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/PaymentItem'
            next_cursor:
              type: string
              nullable: true
            has_more:
              type: boolean
    PaymentItem:
      type: object
      properties:
        payment_uuid:
          type: string
          format: uuid
        amount:
          type: string
        currency:
          type: string
        status:
          $ref: '#/components/schemas/PaymentStatus'
        created_at:
          type: string
          format: date-time
...
    PaymentStatus:
      type: string
      enum:
        - pending
        - waiting
        - confirming
        - confirmed
        - sending
        - partially_paid
        - finished
        - failed
        - expired
        - refunded
      description: |
        Payment status values:
        - `pending`: Payment created, waiting for initialization
        - `waiting`: Waiting for crypto deposit
        - `confirming`: Transaction detected, confirming on blockchain
        - `confirmed`: Transaction confirmed
        - `sending`: Sending to platform wallet
        - `partially_paid`: Partial amount received
        - `finished`: Payment completed successfully
        - `failed`: Payment failed
        - `expired`: Payment expired before completion
        - `refunded`: Payment was refunded
...
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          description: Error code
          example: 5001
        message:
          type: string
          description: Human-readable error message
          example: "payment not found"
        details:
          type: object
          description: Additional error details
          additionalProperties: true

    # Error codes reference
...